chore(welcome): improve structure of welcome page
